requirements.txt
# --- Core Orchestration ---
langchain==1.2.10
langchain-community==0.3.10
langchain-core==0.3.15

# --- Model Integrations ---
groq==1.0.0
langchain-groq==1.1.2
langchain-huggingface==1.2.0

# --- Vector Store & Processing ---
faiss-cpu==1.13.2
sentence-transformers==3.3.0
beautifulsoup4==4.12.3

# --- Utilities ---
python-dotenv==1.0.1
